ADR-796: Declare Duty Suspended Deliveries Outside UK Page feature implemented
