+++ src/test/resources/features/alcoholToDeclare/AlcoholDutyReturns.feature
@@ -1,4 +1,4 @@
-@Test @ZAP @@a11y
+@Test @ZAP @@a11y @AlcoholToDeclare
 Feature: Alcohol Duty Returns Journey
 
   Background: : Common Steps - Alcohol Duty Returns Journey
@@ -35,23 +35,3 @@ Feature: Alcohol Duty Returns Journey
     When I enter "100" on "Product Volume Page"
     And I click save and continue button on "Product Volume Page"
 #   Then I am presented with the "Pure Alcohol Page"
-
-  Scenario: Alcohol Duty Returns Journey - Declare Duty Suspended Deliveries Page - With option 'Yes'
-    Then I am presented with the "Product Name Page"
-    When I enter "TestName" on "Product Name Page"
-    And I click save and continue button on "Product Name Page"
-    When I enter redirect url for "Declare Duty Suspended Deliveries Page"
-    Then I am presented with the "Declare Duty Suspended Deliveries Page"
-    When I select radio button "Yes" on "Declare Duty Suspended Deliveries Page"
-    And I click save and continue button on "Declare Duty Suspended Deliveries Page"
-#   Then I am presented with the "DSD Guidance Page"
-
-  Scenario: Alcohol Duty Returns Journey - Declare Duty Suspended Deliveries Page - With option 'No'
-    Then I am presented with the "Product Name Page"
-    When I enter "TestName" on "Product Name Page"
-    And I click save and continue button on "Product Name Page"
-    When I enter redirect url for "Declare Duty Suspended Deliveries Page"
-    Then I am presented with the "Declare Duty Suspended Deliveries Page"
-    When I select radio button "No" on "Declare Duty Suspended Deliveries Page"
-    And I click save and continue button on "Declare Duty Suspended Deliveries Page"
-#   Then I am presented with the "Task List Page"
