+++ src/test/resources/features/dutySuspendedDeliveries/DutySuspendedDeliveries.feature
@@ -1,0 +1,37 @@
+@Test @ZAP @@a11y @DutySuspendedDeliveries
+Feature: Alcohol Duty Return - Duty Suspended Deliveries
+
+  Background: : Common Steps - Duty Suspended Deliveries
+    Given I navigate to the "Auth Login Stub Page"
+    When I enter redirectURL on "Auth Login Stub Page"
+    And I click continue on "Auth Login Stub Page"
+
+  Scenario: Alcohol Duty Returns - Declare Duty Suspended Deliveries Page - With option 'Yes'
+    Then I am presented with the "Product Name Page"
+    When I enter "TestName" on "Product Name Page"
+    And I click save and continue button on "Product Name Page"
+    When I enter redirect url for "Declare Duty Suspended Deliveries Page"
+    Then I am presented with the "Declare Duty Suspended Deliveries Page"
+    When I select radio button "Yes" on "Declare Duty Suspended Deliveries Page"
+    And I click save and continue button on "Declare Duty Suspended Deliveries Page"
+#   Then I am presented with the "DSD Guidance Page"
+
+  Scenario: Alcohol Duty Returns - Declare Duty Suspended Deliveries Page - With option 'No'
+    Then I am presented with the "Product Name Page"
+    When I enter "TestName" on "Product Name Page"
+    And I click save and continue button on "Product Name Page"
+    When I enter redirect url for "Declare Duty Suspended Deliveries Page"
+    Then I am presented with the "Declare Duty Suspended Deliveries Page"
+    When I select radio button "No" on "Declare Duty Suspended Deliveries Page"
+    And I click save and continue button on "Declare Duty Suspended Deliveries Page"
+#   Then I am presented with the "Task List Page"
+
+  Scenario: Alcohol Duty Returns - Declare Duty Suspended Deliveries Outside UK Page
+    Then I am presented with the "Product Name Page"
+    When I enter "TestName" on "Product Name Page"
+    And I click save and continue button on "Product Name Page"
+    When I enter redirect url for "Declare Duty Suspended Deliveries Outside UK Page"
+    Then I am presented with the "Declare Duty Suspended Deliveries Outside UK Page"
+    When I enter "150" on "Declare Duty Suspended Deliveries Outside UK Page"
+    And I click save and continue button on "Declare Duty Suspended Deliveries Page"
+#   Then I am presented with the "Declare Duty Suspended Deliveries Within UK Page"
