+++ src/test/scala/uk/gov/hmrc/alcoholDuty/cucumber/stepdefs/BaseStepDef.scala
@@ -89,6 +89,8 @@ trait BaseStepDef
         driver.get(TestConfiguration.url("alcohol-duty-returns-frontend") + "/productVolume")
       case "Declare Duty Suspended Deliveries Page" =>
         driver.get(TestConfiguration.url("alcohol-duty-returns-frontend") + "/declareDutySuspendedDeliveriesQuestion")
+      case "Declare Duty Suspended Deliveries Outside UK Page" =>
+        driver.get(TestConfiguration.url("alcohol-duty-returns-frontend") + "/declareDutySuspendedDeliveriesOutsideUk")
     }
   }
 }
