+++ src/test/scala/uk/gov/hmrc/alcoholDuty/pages/alcoholDuty/alcoholToDeclare/SmallReliefProducerPage.scala
@@ -14,30 +14,30 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.alcoholDuty.pages.alcoholDuty
+package uk.gov.hmrc.alcoholDuty.pages.alcoholDuty.alcoholToDeclare
 
 import uk.gov.hmrc.alcoholDuty.conf.TestConfiguration
 import uk.gov.hmrc.alcoholDuty.pages.BasePage
 
-object DeclareDutySuspendedDeliveriesPage extends BasePage {
+object SmallReliefProducerPage extends BasePage {
 
-  override val url: String = TestConfiguration.url("alcohol-duty-returns-frontend") + "/declareDutySuspendedDeliveriesQuestion"
-  override val title       = "Do you need to declare alcohol you delivered or received duty suspended?"
+  override val url: String = TestConfiguration.url("alcohol-duty-returns-frontend") + "/smallProducerReliefQuestion"
+  override val title = "Is this product eligible for Small Producer Relief?"
 
   override def expectedPageErrorTitle: Option[String] = Some(
-    "Error: Do you need to declare alcohol you delivered or received duty suspended? - Manage your Alcohol Duty - GOV.UK"
+    "Error: Is this product eligible for Small Producer Relief? - Manage your Alcohol Duty - GOV.UK"
   )
 
   override def expectedPageTitle: Option[String] = Some(
-    "Do you need to declare alcohol you delivered or received duty suspended? - Manage your Alcohol Duty - GOV.UK"
+    "Is this product eligible for Small Producer Relief? - Manage your Alcohol Duty - GOV.UK"
   )
 
-  override def expectedPageHeader: Option[String] = Some("Do you need to declare alcohol you delivered or received duty suspended?")
+  override def expectedPageHeader: Option[String] = Some("Is this product eligible for Small Producer Relief?")
 
   override def clickRadioButton(text: String): Unit = {
     text match {
-      case "Yes" => click on cssSelector("#declare-duty-suspended-deliveries-input")
-      case "No" => click on cssSelector("#declare-duty-suspended-deliveries-input-no")
+      case "Yes" => click on cssSelector("#small-producer-relief-input")
+      case "No" => click on cssSelector("#small-producer-relief-input-no")
     }
   }
 }
